% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FX_functions.R
\name{import_fx_trade_volume}
\alias{import_fx_trade_volume}
\title{This function imports trade volume in fx}
\usage{
import_fx_trade_volume()
}
\description{
This function imports trade volume in fx
}
\details{
The files are in different formats
\itemize{
\item In the period 2001 - 2008 the data format includes data
about the acitivity of three groups : domestic banks,
foreign financial institutions and other clients
}
}
